Move gtag script to a Component
