+++ src/pages/index.tsx
@@ -1,5 +1,6 @@
 import { GetServerSideProps } from 'next';
-import Script from 'next/script';
+
+import { GtagScript } from '~/components/globals/GtagScript';
 
 import Head from '~/components/globals/Head';
 import CategoriesSlider from '~/components/home/CategoriesSlider';
@@ -11,7 +12,7 @@ import keywords from '~/data/keywords.json';
 
 import Store from '~/lib/Store';
 
-const { NEXT_PUBLIC_BASE_URL, GTM_ID } = process.env;
+const { NEXT_PUBLIC_BASE_URL } = process.env;
 
 type HomeProps = {
   products: Product[];
@@ -30,22 +31,7 @@ const Home = ({ products, categories, promotion, firstProductPromotion }: HomePr
         keywords={keywords.home}
         url={`${NEXT_PUBLIC_BASE_URL}/`}
       />
-      {/* Global site tag (gtag.js) - Google Analytics */}
-      <Script
-        async
-        src={`https://www.googletagmanager.com/gtag/js?id=${GTM_ID}`}
-        strategy="afterInteractive"
-      />
-      <Script id="google-analytics" strategy="afterInteractive">
-        {`
-          window.dataLayer = window.dataLayer || [];
-          function gtag(){window.dataLayer.push(arguments);}
-          gtag('js', new Date());
-          
-          gtag('config', '${GTM_ID}');
-        `}
-      </Script>
-
+      <GtagScript />
       <HomeHero />
       <CategoriesSlider categories={categories} />
       {/* TODO: Update call for specific filter */}
